수정 Category 작업 구조 생성, 패키지 구조 수정
application -> admin , user, model 로 구분지음

+++ src/main/java/com/karljeong/fourtysix/database/repository/TbArticleDiaryRepository.java
@@ -1,16 +1,15 @@
 package com.karljeong.fourtysix.database.repository;
 
+import java.math.BigInteger;
 import java.util.List;
-
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
-
 import com.karljeong.fourtysix.database.entity.TbArticleDiary;
 
 @RepositoryRestResource(collectionResourceRel = "articleEntity", path = "tbArticleDiary")
-public interface TbArticleDiaryRepository extends PagingAndSortingRepository<TbArticleDiary, Long>, JpaSpecificationExecutor<TbArticleDiary> {
+public interface TbArticleDiaryRepository extends PagingAndSortingRepository<TbArticleDiary, BigInteger>, JpaSpecificationExecutor<TbArticleDiary> {
 
     List<TbArticleDiary> findByArticleTitle(@Param("articleTitle") String articleTitle);
